requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cssselect==1.1.0
cycler==0.11.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
fonttools==4.37.4
idna==3.4
jieba3k==0.35.1
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.1
newspaper3k==0.2.8
nltk==3.7
numpy==1.23.4
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.64.1
urllib3==1.26.12
wordcloud==1.8.2.2
